+++ src/all/background_page/controller/move/moveFolderController.js
@@ -11,6 +11,8 @@
  * @link          https://www.passbolt.com Passbolt(tm)
  * @since         2.13.0
  */
+const Worker = require('../../model/worker');
+
 const {Crypto} = require('../../model/crypto');
 const {Keyring} = require('../../model/keyring');
 const {Share} = require('../../model/share');
@@ -226,6 +228,16 @@ class MoveFolderController {
    * @returns {Promise<void>}
    */
   async share() {
+    if (this.foldersChanges.length || this.resourcesChanges.length) {
+      // Some permission changes are possible, ask user what strategy to adopt
+      // This also serves as a confirmation prior to a typically sensitive operation
+      const reactWorker = this.getReactWorker();
+      const strategy = await reactWorker.port.request('passbolt.folders.move-strategy.request');
+      if (strategy.moveOption === 'keep') {
+        return;
+      }
+    }
+
     // Update folders permissions
     if (this.foldersChanges.length) {
       let folders = new FoldersCollection([this.folder]);
@@ -264,6 +276,19 @@ class MoveFolderController {
     this.passphrase = null;
     this.privateKey = null;
   }
+
+  /**
+   * The treatment of the requests coming from any legacy worker should be delegated to the new
+   * react application.
+   * @return {Worker}
+   */
+  getReactWorker() {
+    if (this.worker.isLegacyWorker()) {
+      return Worker.get('ReactApp', this.worker.tab.id);
+    }
+    return this.worker;
+  };
+
 }
 
 exports.MoveFolderController = MoveFolderController;